Moved tests for CatalogConfigResolver to a separate file
