+++ tests/io/test_kedro_data_catalog.py
@@ -11,7 +11,6 @@ from pandas.testing import assert_frame_equal
 
 from kedro.io import (
     CachedDataset,
-    CatalogConfigResolver,
     DatasetAlreadyExistsError,
     DatasetError,
     DatasetNotFoundError,
@@ -494,41 +493,6 @@ class TestKedroDataCatalog:
             pattern = "Unable to find credentials 'other_credentials'"
             with pytest.raises(KeyError, match=pattern):
                 KedroDataCatalog.from_config(**correct_config_with_nested_creds)
-
-        def test_unresolve_config_credentials(self, correct_config):
-            """Test unresolve dataset credentials to original format."""
-            config = correct_config["catalog"]
-            credentials = correct_config["credentials"]
-            resolved_configs = CatalogConfigResolver._resolve_config_credentials(
-                config, credentials
-            )
-
-            unresolved_config, unresolved_credentials = (
-                CatalogConfigResolver.unresolve_config_credentials(
-                    cred_name="s3", ds_config=resolved_configs
-                )
-            )
-            assert config == unresolved_config
-            assert credentials == unresolved_credentials
-
-        def test_unresolve_config_credentials_two_keys(self, correct_config):
-            """Test unresolve dataset credentials to original format when two credentials keys provided."""
-            config = correct_config["catalog"]
-            credentials = correct_config["credentials"]
-
-            resolved_configs = CatalogConfigResolver._resolve_config_credentials(
-                config, credentials
-            )
-            resolved_configs["cars"]["metadata"] = {"credentials": {}}
-
-            unresolved_config, unresolved_credentials = (
-                CatalogConfigResolver.unresolve_config_credentials(
-                    cred_name="s3", ds_config=resolved_configs
-                )
-            )
-            unresolved_config["cars"].pop("metadata")
-            assert config == unresolved_config
-            assert credentials == unresolved_credentials
 
         def test_missing_dependency(self, correct_config, mocker):
             """Test that dependency is missing."""
